h67/hasker/requirements.txt
asgiref==3.7.0
Django
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow
psycopg2==2.9.9
PyJWT==1.7.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions
tzdata==2023.3